refactored into Model-service-repository
