+++ Bookings/Booking.cs
@@ -11,11 +11,6 @@ public class Booking
     public DateTime BookingDate { get; set; } = DateTime.Now;
     public BookingStatus Status { get; set; } = BookingStatus.Pending;
 
-    public static string ToCsv(Booking booking)
-    {
-        return $"{booking.Id},{booking.Flight.FlightNumber},{booking.User.Id},{booking.BookingDate},{booking.Status}";
-    }
-
     public static string header = "bookingId,flightNumber,userId,bookingDate,status";
 
     public Booking(int id, User user, Flight flight)
@@ -34,31 +29,6 @@ public class Booking
         Status = status;
     }
 
-    public static Booking FromCsv(string csv)
-    {
-        string[] values = csv.Split(',');
-
-        #region early fall check
-        if (int.TryParse(values[0], out int bookingId))
-            throw new Exception("Invalid Booking ID");
-
-        if (int.TryParse(values[1], out int flightNumber))
-            throw new Exception("Invalid Flight Number");
-
-        if (int.TryParse(values[2], out int userId))
-            throw new Exception("Invalid User ID");
-
-        if (DateTime.TryParse(values[3], out DateTime bookingDate))
-            throw new Exception("Invalid Booking Date");
-
-        if (Enum.TryParse(values[4], out BookingStatus status))
-            throw new Exception("Invalid Booking Status");
-        #endregion
-
-        Flight? flight = Flight.GetFlightByNumber(flightNumber) ?? throw new Exception("Flight Not Found");
-        Booking booking = new(bookingId, User.GetUserById(userId), flight, bookingDate, status);
-        return booking;
-    }
     public override string ToString()
     {
         return $"Booking ID: {Id}\nFlight: {Flight.FlightNumber}\nUser: {User.Name}\nBooking Date: {BookingDate}\nStatus: {Status}";
